requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
altair
annotated-types
anyio
attrs
blinker
cachetools
certifi
cffi
charset-normalizer
click
cryptography
dataclasses-json
distro
faiss-cpu
frozenlist
gitdb
GitPython
groq
h11
httpcore
httpx
httpx-sse
idna
Jinja2
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
langchain
langchain-community
langchain-core
groq
langchain-text-splitters
langsmith
markdown-it-py
MarkupSafe
marshmallow
mdurl
multidict
mypy-extensions
narwhals
numpy
ollama
orjson
packaging
pandas
pdfminer.six
pdfplumber
pillow
propcache
protobuf
pyarrow
pycparser
pydantic
pydantic-settings
pydantic_core
pydeck
Pygments
pypdfium2
python-dateutil
python-dotenv
pytz
PyYAML
referencing
requests
requests-toolbelt
rich
rpds-py
six
smmap
sniffio
SQLAlchemy
streamlit
tenacity
toml
tornado
typing-inspect
typing_extensions
tzdata
urllib3
yarl
zstandard
rank-bm25
langdetect
pyvis